requirements.txt
ollama 
numpy 
pandas 
tqdm 
faiss-cpu 
rank-bm25
networkx
matplotlib